--- a/supabase/migrations/20250825025822_create_kv_table.sql
+++ b/supabase/migrations/20250825025822_create_kv_table.sql
@@ -1,0 +1,7 @@
+-- Migration: Create KV store table (existing remote migration)
+-- This file matches the remote migration that already exists
+
+CREATE TABLE kv_store_171cbf6f (
+  key TEXT NOT NULL PRIMARY KEY,
+  value JSONB NOT NULL
+);