… MIGRATION_README.md
… PRODUCTION_DEPLOYMENT.md
… package.json
… src/components/MatchHistory.tsx
… src/components/admin/GroupManagement.test.tsx
… src/components/admin/MigrationControl.tsx
… src/components/dashboard/Dashboard.tsx
… src/hooks/useAdaptiveQueries.ts
… src/hooks/useHybridQueries.ts
… src/hooks/useMutations.ts
… src/hooks/useQueries.ts
… src/tests/utils/test-utils.tsx
… src/types/index.ts
… supabase/functions/_shared/data-migration.tsx
… supabase/functions/_shared/kv_store.tsx
… supabase/functions/_shared/kv_store_deprecated.tsx
… supabase/functions/_shared/relational-routes.tsx
… supabase/functions/api/index.tsx
… supabase/functions/data-migration/index.tsx
… supabase/migrations/20241201120000_add_relational_tables.sql
… supabase/migrations/20241201120100_migrate_kv_data.sql
… supabase/migrations/20241201120200_fix_match_players_schema.sql
… supabase/migrations/20241201120300_cleanup_kv_store.sql